ToC update + new files
Most of the new files are empty right now, and in this state, the
existing code almost certainly will not work properly (in the middle of
moving stuff out of init (which was core) and into normalStagger (which
is currently the only functioning part of the addon). Will fix later
tonight, but no time right now.